requirements.txt
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
google-auth==1.30.0
greenlet==1.0.0
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.25.1
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.13
typing-extensions==3.10.0.0
urllib3==1.26.4
waitress==2.0.0
Werkzeug==1.0.1
zipp==3.4.1
